requirements.txt
Flask
Flask-Session
SQLAlchemy
psycopg2
psycopg2-binary
python-dotenv
gunicorn
requests
